+++ api/serializer/machine.py
@@ -1,0 +1,10 @@
+from rest_framework import serializers
+from api.models import Machine
+
+
+class MachineSerializer(serializers.ModelSerializer):
+    """机器信息序列化器"""
+
+    class Meta:
+        model = Machine
+        fields = "__all__"
